Emotion Detector/requirements.txt
opencv2
deepface
datetime
tf-keras